auth: Create Firebase project (#22)
This PR adds the Firebase dependencies and initializes the firebase and firestore global instance for the app. This change can break local builds if a .env.local file is not created in the root of the project.

Let me know and I will send over the .env.local details. Additionally, if you would like access to the Firebase project, send me an email I can use to add yall as a member to the project!